feat(tenjiku): add Tenjiku Shogi (16x16) validated vs HaChu source tables (#402) (#461)
* feat(tenjiku): add 16x16 geometry, roles and Tenjiku Shogi rule layer (#402)

* test(tenjiku): perft pins, HaChu harness driver and variant docs (#402)

* fix(tenjiku): satisfy fmt/clippy/doc gates and u16 SQUARES ripple (#402)

* fix(tenjiku): fifth-tier tokens in notation module and role round-trip tests (#402)

* test(tenjiku): bump registered-variant count to 68 (#402)

* fix(lion): emit the jitto pass once per side so SAN round-trips (#402)

+++ src/geometry/square.rs
@@ -83,7 +83,7 @@ impl<G: Geometry> Square<G> {
     #[must_use]
     #[inline]
     pub const fn new(index: u8) -> Square<G> {
-        assert!(index < G::SQUARES, "square index out of range");
+        assert!((index as u16) < G::SQUARES, "square index out of range");
         Square(index, PhantomData)
     }
 
@@ -92,7 +92,7 @@ impl<G: Geometry> Square<G> {
     #[must_use]
     #[inline]
     pub const fn try_new(index: u8) -> Option<Square<G>> {
-        if index < G::SQUARES {
+        if (index as u16) < G::SQUARES {
             Some(Square(index, PhantomData))
         } else {
             None
